refactor: move mapping functions from data-service.js to new file
